Penambahan Fitur

+++ Code/Lists.hpp
@@ -39,5 +39,5 @@ namespace big::Lists
 		"Join Crew Session", "Join SCTV", "Leave GTA Online"
 	};
 
-	inline std::size_t session_list_pos = 0;
+	inline std::int32_t session_list_pos = 0;
 }
